refactors into higher level collections
