--- api/models/Sheet.js
@@ -1,7 +1,0 @@
-const mongoose = require('mongoose');
-
-const sheetSchema = new mongoose.Schema({
-	items: [{ type: mongoose.Schema.ObjectId, ref: 'Item' }]
-});
-
-module.exports = mongoose.model('Sheet', sheetSchema);
