+++ api/models/Skill.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
 
-const skillsArraySchema = new mongoose.Schema({    
+module.exports = new mongoose.Schema({    
     name: {
         type: String
     },
@@ -9,9 +9,5 @@ const skillsArraySchema = new mongoose.Schema({
     }
 });
 
-const skillsSchema = new mongoose.Schema({
-    game: String,
-    skills: [skillsArraySchema]
-});
+ 
 
-module.exports = mongoose.model('Skill', skillsSchema);
