+++ api/models/User.js
@@ -8,8 +8,7 @@ const userSchema = new mongoose.Schema({
 	password: {
 		type: String,
 		required: 'Please enter a password'
-	},
-	sheets: [{ type: mongoose.Schema.ObjectId, ref: 'Sheet' }]
+	},	
 });
 
 module.exports = mongoose.model('User', userSchema);
