+++ api/resolvers/index.js
@@ -1,6 +1,5 @@
 const userResolver = require('./userResolver');
-const itemResolver = require('./itemResolver');
-const skillResolver = require('./skillResolver');
-const raceResolver = require('./raceResolver');
+const savageWorldsResolver = require('./savageWorldsResolver');
 
-module.exports = [userResolver, itemResolver, skillResolver, raceResolver];
+
+module.exports = [userResolver, savageWorldsResolver];
